Add files via upload
The multiple Regression python code works to predict the calories a person requires taking his Age and Gender as the variables Calories depend on. We take 70% of the dataset to train the model and rest 30% to test it. y_pred has the predicted values which can be compared to y_test, the actual values. A graph for the same will be uploaded soon. And it's efficiency will be improved.